ai-rag-service/requirements.txt
chromadb==0.5.4
sentence-transformers==3.3.1
python-dotenv==1.0.1
google-generativeai==0.8.3
Flask==3.0.3
flask-cors==5.0.0
requests==2.31.0 